EPMLABSBRN-923 replace methods from AdminController to AdminControllerV2
