+++ src/main/kotlin/com/epam/brn/controller/AdminController.kt
@@ -75,15 +75,6 @@ class AdminController(
     ) = ResponseEntity.ok()
         .body(BaseResponseDto(data = studyHistoryService.getHistories(userId, from, to)))
 
-    @GetMapping(value = ["/histories"], params = ["version=2"])
-    @ApiOperation("Get user's study histories for period")
-    fun getHistories(
-        @RequestParam("userId", required = true) userId: Long,
-        @RequestParam("from", required = true) from: LocalDateTime,
-        @RequestParam("to", required = true) to: LocalDateTime
-    ) = ResponseEntity.ok()
-        .body(BaseResponseDto(data = studyHistoryService.getHistories(userId, from, to)))
-
     @GetMapping("/monthHistories")
     @ApiOperation("Get month user's study histories")
     fun getMonthHistories(
@@ -110,17 +101,6 @@ class AdminController(
         return ResponseEntity.ok().body(BaseSingleObjectResponseDto(data = response))
     }
 
-    @GetMapping(value = ["/study/week"], params = ["version=2"])
-    @ApiOperation("Get user's weekly statistic for the period")
-    fun getUserWeeklyStatistic(
-        @RequestParam(name = "from", required = true) from: LocalDateTime,
-        @RequestParam(name = "to", required = true) to: LocalDateTime,
-        @RequestParam(name = "userId", required = true) userId: Long
-    ): ResponseEntity<BaseSingleObjectResponseDto> {
-        val result = userDayStatisticService.getStatisticForPeriod(from, to, userId)
-        return ResponseEntity.ok().body(BaseSingleObjectResponseDto(data = result))
-    }
-
     @GetMapping("/study/year")
     @ApiOperation("Get user's yearly statistic for the period")
     @Deprecated(message = "Use the same method with LocalDateTime as the dates type instead")
@@ -136,17 +116,6 @@ class AdminController(
             it.toDto()
         }
         return ResponseEntity.ok().body(BaseSingleObjectResponseDto(data = response))
-    }
-
-    @GetMapping(value = ["/study/year"], params = ["version=2"])
-    @ApiOperation("Get user's yearly statistic for the period")
-    fun getUserYearlyStatistic(
-        @RequestParam(name = "from", required = true) from: LocalDateTime,
-        @RequestParam(name = "to", required = true) to: LocalDateTime,
-        @RequestParam(name = "userId", required = true) userId: Long
-    ): ResponseEntity<BaseSingleObjectResponseDto> {
-        val result = userMonthStatisticService.getStatisticForPeriod(from, to, userId)
-        return ResponseEntity.ok().body(BaseSingleObjectResponseDto(data = result))
     }
 
     @PostMapping("/loadTasksFile")
