+++ src/main/kotlin/com/epam/brn/controller/AdminControllerV2.kt
@@ -1,0 +1,57 @@
+package com.epam.brn.controller
+
+import com.epam.brn.dto.BaseResponseDto
+import com.epam.brn.dto.BaseSingleObjectResponseDto
+import com.epam.brn.dto.statistic.DayStudyStatistic
+import com.epam.brn.dto.statistic.MonthStudyStatistic
+import com.epam.brn.service.StudyHistoryService
+import com.epam.brn.service.statistic.UserPeriodStatisticService
+import io.swagger.annotations.Api
+import io.swagger.annotations.ApiOperation
+import org.springframework.http.ResponseEntity
+import org.springframework.web.bind.annotation.GetMapping
+import org.springframework.web.bind.annotation.RequestMapping
+import org.springframework.web.bind.annotation.RequestParam
+import org.springframework.web.bind.annotation.RestController
+import java.time.LocalDateTime
+
+@RestController
+@RequestMapping("/v2/admin")
+@Api(value = "/admin", description = "Contains actions for admin")
+class AdminControllerV2(
+    private val studyHistoryService: StudyHistoryService,
+    private val userDayStatisticService: UserPeriodStatisticService<DayStudyStatistic>,
+    private val userMonthStatisticService: UserPeriodStatisticService<MonthStudyStatistic>,
+) {
+
+    @GetMapping("/histories")
+    @ApiOperation("Get user's study histories for period")
+    fun getHistories(
+        @RequestParam("userId", required = true) userId: Long,
+        @RequestParam("from", required = true) from: LocalDateTime,
+        @RequestParam("to", required = true) to: LocalDateTime
+    ) = ResponseEntity.ok()
+        .body(BaseResponseDto(data = studyHistoryService.getHistories(userId, from, to)))
+
+    @GetMapping("/study/week")
+    @ApiOperation("Get user's weekly statistic for the period")
+    fun getUserWeeklyStatistic(
+        @RequestParam(name = "from", required = true) from: LocalDateTime,
+        @RequestParam(name = "to", required = true) to: LocalDateTime,
+        @RequestParam(name = "userId", required = true) userId: Long
+    ): ResponseEntity<BaseSingleObjectResponseDto> {
+        val result = userDayStatisticService.getStatisticForPeriod(from, to, userId)
+        return ResponseEntity.ok().body(BaseSingleObjectResponseDto(data = result))
+    }
+
+    @GetMapping("/study/year")
+    @ApiOperation("Get user's yearly statistic for the period")
+    fun getUserYearlyStatistic(
+        @RequestParam(name = "from", required = true) from: LocalDateTime,
+        @RequestParam(name = "to", required = true) to: LocalDateTime,
+        @RequestParam(name = "userId", required = true) userId: Long
+    ): ResponseEntity<BaseSingleObjectResponseDto> {
+        val result = userMonthStatisticService.getStatisticForPeriod(from, to, userId)
+        return ResponseEntity.ok().body(BaseSingleObjectResponseDto(data = result))
+    }
+}
