+++ src/test/kotlin/com/epam/brn/controller/AdminControllerTest.kt
@@ -43,7 +43,6 @@ import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
 import org.springframework.data.domain.Pageable
 import org.springframework.web.multipart.MultipartFile
-import java.time.LocalDateTime
 
 @ExtendWith(MockKExtension::class)
 internal class AdminControllerTest {
@@ -144,22 +143,6 @@ internal class AdminControllerTest {
     }
 
     @Test
-    fun `getHistories should return histories`() {
-        // GIVEN
-        val userId = 1L
-        val date = LocalDateTime.now()
-        every { studyHistoryService.getHistories(userId, date, date) } returns listOf(studyHistoryDto)
-
-        // WHEN
-        val histories = adminController.getHistories(userId, date, date)
-
-        // THEN
-        verify(exactly = 1) { studyHistoryService.getHistories(userId, date, date) }
-        histories.statusCodeValue shouldBe HttpStatus.SC_OK
-        histories.body!!.data shouldBe listOf(studyHistoryDto)
-    }
-
-    @Test
     fun `getMonthHistories should return month histories`() {
         // GIVEN
         val userId = 1L
@@ -219,44 +202,6 @@ internal class AdminControllerTest {
         verify(exactly = 1) { resourceService.updateDescription(id, description) }
         updated.statusCodeValue shouldBe HttpStatus.SC_OK
         updated.body!!.data shouldBe resourceDto
-    }
-
-    @Test
-    fun `getUserWeeklyStatistic should return weekly statistic`() {
-        // GIVEN
-        val userId = 1L
-        val date = LocalDateTime.now()
-        every { userDayStatisticService.getStatisticForPeriod(date, date, userId) } returns listOf(dayStudyStatistic)
-
-        // WHEN
-        val userWeeklyStatistic = adminController.getUserWeeklyStatistic(date, date, userId)
-
-        // THEN
-        verify(exactly = 1) { userDayStatisticService.getStatisticForPeriod(date, date, userId) }
-        userWeeklyStatistic.statusCodeValue shouldBe HttpStatus.SC_OK
-        userWeeklyStatistic.body!!.data shouldBe listOf(dayStudyStatistic)
-    }
-
-    @Test
-    fun `getUserYearlyStatistic should return yearly statistic`() {
-        // GIVEN
-        val userId = 1L
-        val date = LocalDateTime.now()
-        every {
-            userMonthStatisticService.getStatisticForPeriod(
-                date,
-                date,
-                userId
-            )
-        } returns listOf(monthStudyStatistic)
-
-        // WHEN
-        val userYearlyStatistic = adminController.getUserYearlyStatistic(date, date, userId)
-
-        // THEN
-        verify(exactly = 1) { userMonthStatisticService.getStatisticForPeriod(date, date, userId) }
-        userYearlyStatistic.statusCodeValue shouldBe HttpStatus.SC_OK
-        userYearlyStatistic.body!!.data shouldBe listOf(monthStudyStatistic)
     }
 
     @Test
